Sentiment Analysis on Reviews
Kozmos, a company focused on producing home textiles and casual wear, aims to increase its sales by analyzing customer reviews and improving product features based on complaints. In line with this goal, sentiment analysis will be applied to the reviews, and the data will be tagged. A classification model will then be built using the tagged data.